% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeSprouting.R
\name{alleffect}
\alias{alleffect}
\title{Computation of the Effects in all NCT partitions}
\usage{
alleffect(output, tree_info, N, W, G, Y, X, Ne, Nel, p, minsize)
}
\arguments{
\item{output}{Desired output of the analysis. if output = "detection" only point estimates
are computed, if output = "estimation" both estimated effects and variances are computed}

\item{tree_info}{An NCT data frame}

\item{N}{Sample size}

\item{W}{N x 1 vector, Individual Treatment}

\item{G}{N x 1 vector, Neighborhood Treatment}

\item{Y}{N x 1 vector, Observed Outcome}

\item{X}{N x K matrix, Observed Covariates Matrix}

\item{Ne}{N x 1 vector, Degree}

\item{Nel}{List of N elements - where N is the sample size -
where each element i contains the IDs of the direct neighbors of unit i}

\item{p}{N x 1 vector, Probability to be assigned to the active individual intervention}

\item{minsize}{minimum number of observations for each level of the joint intervention
to be required in the leafs}

\item{Peff}{4 dimensional vector containing the estimated effects in the
whole population}
}
\value{
A data frame reporting the results of the Network Causal Trees. #' Specifically the OF variable includes the value of the OF in the corresponding partition, the NOBS_TR column
includes the number of obs belonging to the training set and to the given partition,
the column FILTER includes the values of the Xs that identify the given partition, the
IDTREE variable identifies the id of the tree ,  the
NUMTREE variable represents the number of trees that identify the given partition,
the NOBS_EST column includes the number of obs belonging
to the estimation set and to the given partition, the columns EFF1000_EST, EFF1101_EST,  EFF1110_EST,
EFF0100_EST reporting the estimated effects in the partitions. If output = "Estimation" the data frame
also include the columns SETAU1000_EST, SETAU1101_EST,  SETAU1110_EST, SETAU0100_EST reporting the estimated standard errors of the effects in the partitions
}
\description{
Computes the estimates in all the partitions identified by the Network Causal Tree
}
