% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_OF_Split.R
\name{compute_OF_Split}
\alias{compute_OF_Split}
\title{Split Objective Function}
\usage{
compute_OF_Split(
  method,
  alpha,
  beta,
  gamma,
  delta,
  W,
  G,
  Y,
  X,
  p,
  Ne,
  Ne_list,
  population_effects,
  total_variance,
  nleafs
)
}
\arguments{
\item{method}{method to compute the Objective function: "singular" for NCT targeted to one single effect;
"composite" for NCT targeted to multiple effects; "penalized" for a OF computed while
considering a single effect only and including a penalization term related to the variance}

\item{alpha}{weight associated to the effect 1000}

\item{beta}{weight associated to the effect 1101}

\item{gamma}{weight associated to the effect 1110}

\item{delta}{weight associated to the effect 0100}

\item{W}{N x 1 vector, Individual Treatment}

\item{G}{N x 1 vector, Neighborhood Treatment}

\item{Y}{N x 1 vector, Observed Outcome}

\item{X}{N x M matrix, Observed Covariate Matrix}

\item{p}{N x 1 vector, Probability to be assigned to the active individual intervention}

\item{Ne}{N x 1 vector, Degree}

\item{Ne_list}{List of N elements - where N is the sample size -
where each element i contains the IDs of the direct neighbors of unit i}

\item{population_effects}{4 dimensional vector containing the estimated effects in the
whole population}

\item{total_variance}{\itemize{
\item to be included if method = "penalized" - whole variance
}}

\item{nleafs}{\itemize{
\item to be included if method = "penalized" - number of leafs
}}
}
\value{
A numeric vector made up by three elements: the first one identifies the
value of the  Objective Function is max , the second one reports the value of the variable that maximizes the  Objective Function,
the third one reports the corresponding variable
}
\description{
Splits the sample where the  Objective Function is maximized
}
