% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_population_effects.R
\name{compute_population_effects}
\alias{compute_population_effects}
\title{Estimated Effects in the whole sample}
\usage{
compute_population_effects(N, W, G, Y, p, Ne)
}
\arguments{
\item{N}{Sample size}

\item{W}{N x 1 vector, Individual Treatment}

\item{G}{N x 1 vector, Neighborhood Treatment}

\item{Y}{N x 1 vector, Observed Outcome}

\item{p}{N x 1 vector, Probability to be assigned to the active individual intervention}

\item{Ne}{N x 1 vector, Degree}
}
\value{
A numeric vector made up by four elements, representing the
estimated effects in the whole population: the first element is the effect 1000,
the second one refers to the effect 1101, the third one refers to 1110 and the
fourth refers to 0100
}
\description{
Computes the four Estimated Effects in the Whole Population
}
