% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covy.R
\name{Covy}
\alias{Covy}
\title{Estimated Covariance of Average Potential Outcomes}
\usage{
Covy(w1, g1, w2, g2, N, Y, W, G, p, Ne, Ne_list)
}
\arguments{
\item{w1}{Individual Treatment level - 1}

\item{g1}{Neighborhood Treatment level - 1}

\item{w2}{Individual Treatment level - 2}

\item{g2}{Neighborhood Treatment level -2}

\item{N}{Sample size}

\item{Y}{N x 1 vector, Observed Outcome}

\item{W}{N x 1 vector, Individual Treatment}

\item{G}{N x 1 vector, Neighborhood Treatment}

\item{p}{N x 1 vector, Probability to be assigned to the active individual intervention}

\item{Ne}{N x 1 vector, Degree}

\item{Ne_list}{List of N elements - where N is the sample size -
where each element i contains the IDs of the direct neighbors of unit i}
}
\value{
A numeric value corresponding to the estimated covariance of the
average potential outcome under the level w1,g1 and w2,g2 of the joint intervention
}
\description{
Computes the estimated covariance of the average potential outcomes
related to two given levels of the joint intervention
\itemize{
\item denoted with (w1,g1) and (w2,g2), respectively -.
}
}
