% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_clustered_networks}
\alias{generate_clustered_networks}
\title{Simulate a Clustered Network Environment}
\usage{
generate_clustered_networks(
  m,
  N,
  method_networks,
  param_er,
  var_homophily_ergm,
  coef_ergm
)
}
\arguments{
\item{m}{Number of clusters}

\item{N}{Number of units}

\item{method_networks}{method to generate the m networks: "ergm" (Exponential Random Graph Models) ,
"er" (Erdos Renyi) ,"sf" (Barabasi-Albert model)
Note: in this function, clusters have the same size, so N should be a multiple of m}

\item{param_er}{If method "er", probability of the ER model}

\item{var_homophily_ergm}{Variable to account for homophily}

\item{coef_ergm}{If method "ergm", coefficients of the ERGM model}
}
\value{
: An adjacency matrix which describes a clustered network environment
}
\description{
Generates a simulated clustered network environment
}
