% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Others.R
\name{genmultnet}
\alias{genmultnet}
\title{Heterogeneous Treatment and Spillover Effects
under Clustered Network Interference}
\usage{
genmultnet(m, N, method, param, varhom, coefergm)
}
\arguments{
\item{m}{Number of clusters}

\item{N}{Number of units}

\item{method}{method to generate the networks: "ergm" (Exponential Random Graph Models) , "er" (Erdos Renyi) ,
"sf" (Barabasi-Albert model)}

\item{param}{If method "er", probability of the ER model}

\item{varhom}{Variable to account for homophily}

\item{coefergm}{If method "ergm", coefficients of the ERGM model}
}
\value{
: An adjacency matrix which describes a clustered network environment
}
\description{
Generates a simulated clustered network environment
}
