% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_partitions_nct.R
\name{identify_partitions_nct}
\alias{identify_partitions_nct}
\title{Identification of Partitions of the Network Causal Tree}
\usage{
identify_partitions_nct(
  method,
  alpha,
  beta,
  gamma,
  delta,
  depth,
  minsize,
  N,
  W,
  G,
  Y,
  X,
  p,
  Ne,
  Ne_list,
  population_effects
)
}
\arguments{
\item{method}{method to compute the Objective function: "singular" for NCT targeted to one single effect;
"composite" for NCT targeted to multiple effects; "penalized" for a OF computed while
considering a single effect only and including a penalization term related to the variance}

\item{alpha}{weight associated to the effect 1000}

\item{beta}{weight associated to the effect 1101}

\item{gamma}{weight associated to the effect 1110}

\item{delta}{weight associated to the effect 0100}

\item{depth}{Depth of the tree}

\item{minsize}{Minimum number of observations for each level of the joint intervention
to be required in the leafs}

\item{N}{Sample size}

\item{W}{N x 1 vector, Individual Treatment}

\item{G}{N x 1 vector, Neighborhood Treatment}

\item{Y}{N x 1 vector, Observed Outcome}

\item{X}{N x M matrix, Observed Covariate Matrix}

\item{p}{N x 1 vector, Probability to be assigned to the active individual intervention}

\item{Ne}{N x 1 vector, Degree}

\item{Ne_list}{List of N elements - where N is the sample size -
where each element i contains the IDs of the direct neighbors of unit i}

\item{population_effects}{4 dimensional vector containing the estimated effects in the
whole population}
}
\value{
A data frame describing the Network Causal Tree. Specifically,
the data frame includes the NODE column identifying the node number, the OF variable
including the value of the  Objective Function in the corresponding partition, the NOBS column
including the number of obs in the given partition, the column FILTER including the
valyes of the Xs to identify the given partition, the column TERMINAL reports the
'role' of the node - parent or leaf -
}
\description{
Identifies the partitions of a Network Causal Tree
}
