% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkCausalTree.R
\name{NetworkCausalTree}
\alias{NetworkCausalTree}
\title{Network Causal Tree}
\usage{
NetworkCausalTree(
  X,
  Y,
  W,
  effect_weights = c(1, 0, 0, 0),
  A = NULL,
  K = NULL,
  p = NULL,
  ratio_disc,
  depth = 3,
  minsize = 10,
  method = "singular",
  output = "estimation"
)
}
\arguments{
\item{X}{N x M Observed Covariates Matrix.}

\item{Y}{N x 1 Observed Outcome vector.}

\item{W}{N x 1 Individual Treatment vector.}

\item{effect_weights}{Treatment Effect weights vector (4 elements):
\itemize{
\item alpha: weight associated to the treatment effect 1000 (effect of the
individual treatment, with the neighborhood treatment set at 0),
\item beta: weight associated to the treatment effect 1101 (effect of the
individual treatment, with the neighborhood treatment set at 1),
\item gamma: weight associated to the spillover effect 1110 (effect of the
neighborhood treatment, with the individual treatment set at 1),
\item delta: weight associated to the spillover effect 0100 (effect of the
neighborhood treatment, with the individual treatment set at 0).
}}

\item{A}{N x N Adjacency matrix.}

\item{K}{N x 1 Cluster Membership vector.}

\item{p}{N x 1 Probability to be assigned to the active individual
intervention vector.}

\item{ratio_disc}{Ratio of clusters to be assigned to the discovery set only.}

\item{depth}{Depth of the tree.}

\item{minsize}{Minimum number of observaztions for each level of the joint
intervention to be required in the leafs}

\item{method}{Method to compute the objective function: "singular" for NCT
targeted to one single effect; "composite" for NCT targeted to multiple
effects; "penalized" for a OF computed while considering a single effect only
and including a penalization term related to the variance within the leafs.}

\item{output}{Desired output of the analysis. if output = "detection" only
point estimates are computed, if output = "estimation" both estimated effects
and variances are computed}
}
\value{
A data.frame describing the obtained Network Causal Trees.
Each row represents a partition (of a specific tree) with 10/14 entries.
Columns summary:
\itemize{
\item \code{NOBS_TR}: number of training observations in the partition,
\item \code{FILTER}: values of the covariates \code{X} that identify the partition,
\item \code{NOBS_EST}: number of estimation observations in the partition,
\item \code{EFF1000_EST}: estimated 1000 effects in the partitions,
\item \code{EFF1101_EST}: estimated 1101 effects in the partitions,
\item \code{EFF1110_EST}: estimated 1110 effects in the partitions,
\item \code{EFF0100_EST}: estimated 0100 effects in the partitions.
Additional columns summary (only if output = "Estimation"):
\item \code{SETAU1000_EST}: estimated std. error of the 1000 effect in the partition,
\item \code{SETAU1101_EST}: estimated std. error of the 1101 effect in the partition,
\item \code{SETAU1110_EST}: estimated std. error of the 1110 effect in the partition,
\item \code{SETAU0100_EST}: estimated std. error of the 0100 effect in the partition.
}
}
\description{
Returns a Network Causal Tree, with the corresponding estimates
}
