% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCT.R
\name{NetworkCausalTrees}
\alias{NetworkCausalTrees}
\title{Network Causal Tree}
\usage{
NetworkCausalTrees(
  effweights,
  A,
  p,
  W,
  Y,
  X,
  M,
  G,
  Ne,
  mdisc,
  mest,
  method,
  minpopfrac = 1,
  depth = 3,
  fracpredictors = 1,
  minsize = 10,
  n_trees = 1,
  output = "estimation"
)
}
\arguments{
\item{effweights}{Vector including the 4 weights alpha, beta, gamma, delta where
\itemize{
\item alpha weight associated to the effect 1000
\item beta weight associated to the effect 1101
\item gamma weight associated to the effect 1110
\item delta weight associated to the effect 0100
}}

\item{A}{adjacency matrix}

\item{p}{N x 1 vector, Probability to be assigned to the active individual intervention}

\item{W}{N x 1 vector, Individual Treatment}

\item{Y}{N x 1 vector, Observed Outcome}

\item{X}{N x K matrix, Observed Covariates Matrix}

\item{M}{N x 1 vector, Cluster Membership}

\item{G}{N x 1 vector, Neighborhood Treatment}

\item{Ne}{N x 1, Degree}

\item{mdisc}{Numeric value, number of clusters to be assigned to the discovery set}

\item{mest}{Numeric value, number of clusters to be assigned to the estimation set only}

\item{method}{method to compute the Objective function: "singular" for NCT targeted to one single effect;
"composite" for NCT targeted to multiple effects; "penalized" for a OF computed while
considering a single effect only and including a penalization term related to the variance}

\item{minpopfrac}{Quote of the discovery set population to be included while sprouting the tree}

\item{depth}{Depth of the tree}

\item{fracpredictors}{Quote of the predictors to be included while sprouting the tree}

\item{minsize}{Minimum number of observaztions for each level of the joint intervention
to be required in the leafs}

\item{output}{Desired output of the analysis. if output = "detection" only point estimates
are computed, if output = "estimation" both estimated effects and variances are computed}

\item{N}{Sample size}

\item{n_tres}{Number of Trees}
}
\value{
A Network Causal Tree - NCT - object. An NCT object is a data frame reporting the
results of the Network Causal Trees, where each tree is characterized by a set of partitions.
Specifically the GOF variable includes the value of the GOF in the corresponding partition, the NOBS_TR column
includes the number of obs belonging to the training set and to the given partition,
the column FILTER includes the values of the Xs that identify the given partition, the
IDTREE variable identifies the id of the tree ,  the
NUMTREE variable represents the number of trees that identify the given partition,
the NOBS_EST column includes the number of obs belonging
to the estimation set and to the given partition, the columns EFF1000_EST, EFF1101_EST,  EFF1110_EST,
EFF0100_EST reporting the estimated effects in the partitions. If output = "Estimation" the data frame
also include the columns SETAU1000_EST, SETAU1101_EST,  SETAU1110_EST, SETAU0100_EST reporting the estimated standard errors of the effects in the partitions
}
\description{
Returns a Network Causal Tree, with the corresponding estimates
}
