% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi.R
\name{pi}
\alias{pi}
\title{Individual Marginal Exposure Probabilities}
\usage{
pi(i, w, g, p, Ne)
}
\arguments{
\item{i}{Unit ID}

\item{w}{Individual Treatment level}

\item{g}{Neighborhood Treatment level}

\item{p}{N x 1 vector, Probability to be assigned to the active individual intervention}

\item{Ne}{N x 1 vector, Degree}
}
\value{
A numeric value bounded between 0 and 1 measuring
the individual marginal probability to be exposed
to the level (w,g) of the joint intervention
}
\description{
Computes the individual marginal probability to be exposed to a given
level of the joint intervention.
}
