% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilities.R
\name{pij}
\alias{pij}
\title{Joint Exposure Probabilities}
\usage{
pij(i, j, wi, wj, gi, gj, Ne, Nel, p)
}
\arguments{
\item{i}{Unit ID}

\item{j}{Unit ID}

\item{wi}{Individual Treatment level}

\item{wj}{Individual Treatment level}

\item{gi}{Neighborhood Treatment level}

\item{gj}{Neighborhood Treatment level}

\item{Ne}{N x 1 vector, Degree}

\item{Nel}{List of N elements - where N is the sample size -
where each element i contains the IDs of the direct neighbors of unit i}

\item{p}{N x 1 vector, Probability to be assigned to the active individual intervention}
}
\value{
A numeric value bounded between 0 and 1 measuring
the individual joint probability of units i and j to be simultaneously
exposed to the level wi,gi and wj,gj of  the joint intervention, respectively
}
\description{
Computes the individual joint probability of a given pair of units
to be simultaneously exposed to certain levels of the joint intervention
}
