% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotNCT.R
\name{plot.NetworkCausalTrees}
\alias{plot.NetworkCausalTrees}
\title{Heterogeneous Treatment and Spillover Effects
under Clustered Network Interference}
\usage{
\method{plot}{NetworkCausalTrees}(
  NCT,
  output,
  vcolor = c("seagreen4", "seagreen1", "lightblue1", "dodgerblue2"),
  vlabelcolor,
  coloreff,
  ewidth,
  elabelcex,
  ecolor,
  elabelfamily,
  elabelcolor,
  vsize,
  vsize2,
  vshape,
  varnames,
  vlabelfont,
  vlabelcex,
  ot = FALSE,
  colleg,
  coltitleleg,
  vframecolor,
  title,
  cex.main,
  col.main,
  adj,
  font.main
)
}
\arguments{
\item{NCT}{A Network Causal Tree object}

\item{output}{Output of the NCT. if output = "detection" only point estimates
are reported, if output = "estimation" both estimated effects and variances are reported}

\item{vcolor}{String Vector of four colors: the first two identify the extremes
of the color palette employed to depict negative values, while the second two identify the extremes
of the color palette employed to depict positive values.
Note that the intensity of the color is related to the Statistic of the
estimated effect in each leaf.
Default values represent negative estimates
with greens and positive estimates with blues}

\item{vlabelcolor}{Vertex label color
Titlw}

\item{coloreff}{Estimated Effect according to which nodes are colored
It can be "1000", "0100","1110","1101"}

\item{ewidth}{Edge width}

\item{elabelcex}{Edge's label cex}

\item{ecolor}{Edges color}

\item{elabelfamily}{Edge label family}

\item{elabelcolor}{Edges' label color
Vertices}

\item{vsize}{Vertex size}

\item{vsize2}{Vertex size - to be set if you want rectangular vertices -}

\item{vshape}{Vertex shape}

\item{varnames}{Vector including the - ordered - names of predictors}

\item{vlabelfont}{Vertex label font}

\item{vlabelcex}{Vertex Label cex}

\item{ot}{TRUE if the NCT object includes an optimal treatment evaluation,
FALSE otherwise (default is FALSE)
Edges}

\item{colleg}{Legend color}

\item{coltitleleg}{Legend's title color}

\item{vframecolor}{Vertex's frame color}

\item{title}{Title}

\item{cex.main}{Title cex}

\item{col.main}{Title color
Legend}

\item{adj}{adj}
}
\description{
Plots the Network Causal Tree
General
}
