% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_NCT}
\alias{plot_NCT}
\title{Plot of a Network Causal Tree}
\usage{
plot_NCT(
  NCT,
  varnames,
  title,
  output = "estimation",
  coloreff = "1000",
  vcolor = c("lightskyblue", "dodgerblue2"),
  vsize = 44,
  vsize2 = 40,
  vshape = "rectangle",
  vframecolor = "black",
  vlabelcolor = "black",
  vlabelcex = 2,
  vlabelfont = 1,
  elabelcex = 1.5,
  elabelfamily = "sans",
  ecolor = "black",
  ewidth = 0.3,
  edge.arrow.size = 0.5,
  elabelcolor = "black",
  colleg = "black",
  coltitleleg = "black",
  font.main = 1,
  cex.main = 1,
  adj = 1,
  col.main = "black"
)
}
\arguments{
\item{NCT}{NetworkCausalTrees results.}

\item{varnames}{(Ordered) covariates names vector.}

\item{title}{Title.}

\item{output}{If "detection" only point estimates are reported, if
"estimation" both estimated effects and variances are reported.}

\item{coloreff}{Estimated Effect according to which nodes are colored
It can be "1000", "0100","1110","1101".}

\item{vcolor}{String Vector of two colors identifying the
extremes of the color palette employed to depict positive estimates of the effect specified in "coloreff"
(Negative values are represented in white).}

\item{vsize}{Vertex size.}

\item{vsize2}{Vertex size (for rectangular vertices).}

\item{vshape}{Vertex shape.}

\item{vframecolor}{Vertex's frame color.}

\item{vlabelcolor}{Vertex label color.
Title}

\item{vlabelcex}{Vertex Label cex.}

\item{vlabelfont}{Vertex label font.}

\item{elabelcex}{Edge's label cex.}

\item{elabelfamily}{Edge label family.}

\item{ecolor}{Edges color.}

\item{ewidth}{Edge width.}

\item{edge.arrow.size}{Edge's arrow size.}

\item{elabelcolor}{Edges' label color.
Vertices}

\item{colleg}{Legend color.}

\item{coltitleleg}{Legend's title color.}

\item{cex.main}{Title cex.}

\item{adj}{adj.}

\item{col.main}{Title color.
Legend}

\item{ot}{TRUE if the results. includes an optimal treatment evaluation,
FALSE otherwise (default: FALSE).
Edges}
}
\value{
A plot representing the estimated Network causal tree. Each node reports - from the top to the bottom
\itemize{
\item i) the estimated effect 10000
ii) the estimated effect 0100
iii) the size of the sub population
Colors provide an intuition on the strength of the effect specified in "coloreff".
}
}
\description{
Plots the Network Causal Tree results.
General
}
