% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Others.R
\name{sharedn}
\alias{sharedn}
\title{Heterogeneous Treatment and Spillover Effects
under Clustered Network Interference}
\usage{
sharedn(i, j, Nel)
}
\arguments{
\item{i}{Unit ID}

\item{j}{Unit ID}

\item{Nel}{List of N elements - where N is the sample size -
where each element i contains the IDs of the direct neighbors of unit i}
}
\value{
A numeric value representing the number of shared neighbors between unit i and j
}
\description{
Computes the number of shared neighbors between unit i and unit j
}
