% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprout_nct.R
\name{sprout_nct}
\alias{sprout_nct}
\title{Generation of the Network Causal Tree}
\usage{
sprout_nct(
  method,
  sampled_clusters,
  depth,
  minsize,
  alpha,
  beta,
  gamma,
  delta,
  N,
  W,
  G,
  Y,
  X,
  K,
  Ne,
  p,
  population_effects,
  Ne_list,
  A = NULL
)
}
\arguments{
\item{method}{method to compute the Objective function: "singular" for NCT targeted to one single effect;
"composite" for NCT targeted to multiple effects; "penalized" for a OF computed while
considering a single effect only and including a penalization term related to the variance}

\item{sampled_clusters}{Clusters assigned to the discovery set}

\item{depth}{depth of the tree}

\item{minsize}{minimum number of observaztions for each level of the joint intervention
to be required in the leafs}

\item{alpha}{weight associated to the effect 1000}

\item{beta}{weight associated to the effect 1101}

\item{gamma}{weight associated to the effect 1110}

\item{delta}{weight associated to the effect 0100}

\item{N}{Sample size}

\item{W}{N x 1 vector, Individual Treatment}

\item{G}{N x 1 vector, Neighborhood Treatment}

\item{Y}{N x 1 vector, Observed Outcome}

\item{X}{N x M matrix, Observed Covariates Matrix}

\item{K}{N x 1 vector, Cluster Membership}

\item{Ne}{N x 1 vector, Degree}

\item{p}{N x 1 vector, Probability to be assigned to the active individual intervention}

\item{population_effects}{4 dimensional vector containing the estimated effects in the
whole population}

\item{Ne_list}{List of N elements - where N is the sample size -
where each element i contains the IDs of the direct neighbors of unit i}

\item{A}{Adjacency matrix (internal use)}
}
\value{
A data frame describing the Network Causal Tree. Specifically,
the data frame includes the NODE column identifying the node number, the OF variable
including the value of the OF in the corresponding partition, the NOBS column
including the number of obs in the given partition, the column FILTER including the
valyes of the Xs to identify the given partition,, the column TERMINAL reports the
'role' of the node - parent or leaf -.
}
\description{
Generation of the Network Causal Tree
}
