% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeSprouting.R
\name{sproutnetctree}
\alias{sproutnetctree}
\title{Generation of the Network Causal Tree}
\usage{
sproutnetctree(
  method,
  minpopfrac,
  fracpredictors,
  sampgroup,
  m,
  alpha,
  beta,
  gamma,
  delta,
  depth,
  minsize,
  N,
  W,
  G,
  Y,
  X,
  M,
  p,
  Ne,
  Peff,
  Nel
)
}
\arguments{
\item{method}{method to compute the Objective function: "singular" for NCT targeted to one single effect;
"composite" for NCT targeted to multiple effects; "penalized" for a OF computed while
considering a single effect only and including a penalization term related to the variance}

\item{minpopfrac}{Quote of the discovery set population to be included while sprouting the tree}

\item{fracpredictors}{Quote of the predictors to be included while sprouting the tree}

\item{sampgroup}{Clusters assigned to the discovery set}

\item{m}{Total number of clusters}

\item{alpha}{weight associated to the effect 1000}

\item{beta}{weight associated to the effect 1101}

\item{gamma}{weight associated to the effect 1110}

\item{delta}{weight associated to the effect 0100}

\item{depth}{depth of the tree}

\item{minsize}{minimum number of observaztions for each level of the joint intervention
to be required in the leafs}

\item{N}{Sample size}

\item{W}{N x 1 vector, Individual Treatment}

\item{G}{N x 1 vector, Neighborhood Treatment}

\item{Y}{N x 1 vector, Observed Outcome}

\item{X}{N x K matrix, Observed Covariates Matrix}

\item{M}{N x 1 vector, Cluster Membership}

\item{p}{N x 1 vector, Probability to be assigned to the active individual intervention}

\item{Ne}{N x 1 vector, Degree}

\item{Peff}{4 dimensional vector containing the estimated effects in the
whole population}

\item{Nel}{List of N elements - where N is the sample size -
where each element i contains the IDs of the direct neighbors of unit i}
}
\value{
A list containing two elements 1- A data frame describing the Network Causal Tree. Specifically,
the data frame includes the NODE column identifying the node number, the OF variable
including the value of the OF in the corresponding partition, the NOBS column
including the number of obs in the given partition, the column FILTER including the
valyes of the Xs to identify the given partition,, the column TERMINAL reports the
'role' of the node - parent or leaf -. 2 - a vector that includes the
predictors employed while sprouting the tree
}
\description{
Sprouts the network causal tree, eventually including a fraction of the initial -discovery-
sample or a fraction of predictors.
}
