% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variances.R
\name{Vartau0100}
\alias{Vartau0100}
\title{Estimated Variance for the effect 0100}
\usage{
Vartau0100(N, Y, W, G, p, Ne, Nel)
}
\arguments{
\item{N}{Sample size}

\item{Y}{N x 1 vector, Observed Outcome}

\item{W}{N x 1 vector, Individual Treatment}

\item{G}{N x 1 vector, Neighborhood Treatment}

\item{p}{N x 1 vector, Probability to be assigned to the active individual intervention}

\item{Ne}{N x 1 vector, Degree}

\item{Nel}{List of N elements - where N is the sample size -
where each element i contains the IDs of the direct neighbors of unit i}
}
\value{
A numeric value corresponding to the estimated variance of the effect
of interest.
}
\description{
Computes the estimated variance of the effect of interest
}
