% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variances.R
\name{Vary}
\alias{Vary}
\title{Heterogeneous Treatment and Spillover Effects
under Clustered Network Interference}
\usage{
Vary(N, w, g, Y, W, G, p, Ne, Nel)
}
\arguments{
\item{N}{Sample size}

\item{w}{Individual Treatment level}

\item{g}{Neighborhood Treatment level}

\item{Y}{N x 1 vector, Observed Outcome}

\item{W}{N x 1 vector, Individual Treatment}

\item{G}{N x 1 vector, Neighborhood Treatment}

\item{p}{Probability to be assigned to the active individual intervention}

\item{Ne}{N x 1 vector, Degree}

\item{Nel}{List of N elements - where N is the sample size -
where each element i contains the IDs of the direct neighbors of unit i}
}
\value{
A numeric value corresponding to the estimated variance of the
average potential outcome under the level w,g of the joint intervention
}
\description{
Computes the estimated variance of the average potential outcome
related to a given level of the joint intervention
}
