% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeSprouting.R
\name{alleffect}
\alias{alleffect}
\title{Computation of the Effects in all NCT partitions}
\usage{
alleffect(output, tree_info, N, W, G, Y, X, Ne, Nel, p, minsize)
}
\arguments{
\item{output}{Desired output of the analysis. if output = "detection" only point estimates
are computed, if output = "estimation" both estimated effects and variances are computed}

\item{tree_info}{An NCT data frame}

\item{N}{Sample size}

\item{W}{N x 1 vector, Individual Treatment}

\item{G}{N x 1 vector, Neighborhood Treatment}

\item{Y}{N x 1 vector, Observed Outcome}

\item{X}{N x K matrix, Observed Covariates Matrix}

\item{Ne}{N x 1 vector, Degree}

\item{Nel}{List of N elements - where N is the sample size -
where each element i contains the IDs of the direct neighbors of unit i}

\item{p}{N x 1 vector, Probability to be assigned to the active individual intervention}

\item{minsize}{minimum number of observations for each level of the joint intervention
to be required in the leafs}

\item{Peff}{4 dimensional vector containing the estimated effects in the
whole population}
}
\value{
A data.frame describing the obtained Network Causal Trees.
Each row represents a partition (of a specific tree) with 10/14 entries.
Columns summary:
\itemize{
\item \code{OF}: value of the OF in the corresponding partition,
\item \code{NOBS_TR}: number of training observations in the partition,
\item \code{FILTER}: values of the covariates \code{X} that identify the partition,
\item \code{IDTREE}: tree ID,
\item \code{NUMTREE}: number of trees identifying the partition,
\item \code{NOBS_EST}: number of estimation observations in the partition,
\item \code{EFF1000_EST}: estimated 1000 effects in the partitions,
\item \code{EFF1101_EST}: estimated 1101 effects in the partitions,
\item \code{EFF1110_EST}: estimated 1110 effects in the partitions,
\item \code{EFF0100_EST}: estimated 0100 effects in the partitions.
Additional columns summary (only if output = "Estimation"):
\item \code{SETAU1000_EST}: estimated std. error of the 1000 effect in the partition,
\item \code{SETAU1101_EST}: estimated std. error of the 1101 effect in the partition,
\item \code{SETAU1110_EST}: estimated std. error of the 1110 effect in the partition,
\item \code{SETAU0100_EST}: estimated std. error of the 0100 effect in the partition.
}
}
\description{
Computes the estimates in all the partitions identified by the Network Causal Tree
}
