% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generator.R
\name{data_generator}
\alias{data_generator}
\title{Synthetic data generator}
\usage{
data_generator(
  N = 2000,
  M = 5,
  k = 40,
  p = rep(0.2, 2000),
  het = TRUE,
  h = 2,
  method_networks = "er",
  param_er = 0.1,
  coef_ergm = NULL,
  var_homophily_ergm = NULL,
  remove_isolates = TRUE
)
}
\arguments{
\item{N}{Sample size (default: 2000).}

\item{M}{Number of binary regressors (default: 5).}

\item{k}{Number of clusters (default: 40).}

\item{p}{N x 1 vector, Probability to be assigned to the active individual
intervention (default: rep(0.2,2000))}

\item{het}{TRUE if the treatment effects 1000 and 1101 are heterogeneous with
respect to the first regressor (h with X1=0, -taui with X0=0), FALSE if
constant (-h) (default: TRUE).}

\item{h}{Absolute value of the treatment effects 1000 and 1101
(default: 2).}

\item{method_networks}{Method to generate the m networks:
"ergm" (Exponential Random Graph Models), "er" (Erdos Renyi), "sf"
(Barabasi-Albert model) (default: "er").
Note: in this function, clusters have the same size, so N should be a multiple of m}

\item{param_er}{Probability of the "er" model, if used (default: 0.2).}

\item{coef_ergm}{Coefficients of the "ergm" model, if used (default: NULL).}

\item{var_homophily_ergm}{Variable to account for homophily in the "ergm"}

\item{remove_isolates}{Logical; remove isolated nodes? (default TRUE)
model (default: NULL).}
}
\value{
A list of synthetic data containing:
\itemize{
\item NxM covariates matrix (\code{X}).
\item Nx1 outcome vector (\code{Y}),
\item Nx1 individual intervention vector (\code{W}),
\item NxN adjacency matrix (\code{A}),
\item Nx1 neighborhood intervention vector (\code{G}),
\item Nx1 group membership vector (\code{K}),
\item Nx1 probability to be assigned to the active individual intervention vector
(\code{p}),
}
}
\description{
Generates Network Causal Tree synthetic data.
}
