% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.grid.unique.R
\name{expand.grid.unique}
\alias{expand.grid.unique}
\title{Generate all combinations of the supplied vectors, without rows characterized
by equal elements}
\usage{
expand.grid.unique(x, y, include.equals = FALSE)
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}

\item{include.equals}{Boolean (dafault: FALSE)}
}
\value{
A data frame with all combinations of the elements of the vectors,
with no rows characterized by equal elements
}
\description{
Create a data frame including all combinations of the supplied vectors or factors,
while omitting rows characterized by equal elements.
}
