% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCT.R
\name{NetworkCausalTrees}
\alias{NetworkCausalTrees}
\title{Network Causal Tree}
\usage{
NetworkCausalTrees(
  X,
  Y,
  W,
  effweights = c(0.25, 0.25, 0.25, 0.25),
  A = NULL,
  G = NULL,
  M = NULL,
  p = NULL,
  mdisc = 25,
  mest = 15,
  minpopfrac = 1,
  fracpredictors = 1,
  n_trees = 1,
  depth = 3,
  minsize = 10,
  method = "singular",
  output = "estimation"
)
}
\arguments{
\item{X}{N x K Observed Covariates Matrix.}

\item{Y}{N x 1 Observed Outcome vector.}

\item{W}{N x 1 Individual Treatment vector.}

\item{effweights}{Vector including the 4 effect weight:
\itemize{
\item alpha weight associated to the effect 1000,
\item beta weight associated to the effect 1101,
\item gamma weight associated to the effect 1110,
\item delta weight associated to the effect 0100.
}}

\item{A}{N x N Adjacency matrix.}

\item{G}{N x 1 Neighborhood Treatment vector.}

\item{M}{N x 1 Cluster Membership vector.}

\item{p}{N x 1 Probability to be assigned to the active individual
intervention vector.}

\item{mdisc}{Number of clusters to be assigned to the discovery set.}

\item{mest}{Number of clusters to be assigned to the estimation set only.}

\item{minpopfrac}{Ratio of the discovery set population to be included while
sprouting the tree.}

\item{fracpredictors}{Quote of the predictors to be included while sprouting
the tree}

\item{n_trees}{Number of Trees.}

\item{depth}{Depth of the tree.}

\item{minsize}{Minimum number of observaztions for each level of the joint
intervention
to be required in the leafs}

\item{method}{Method to compute the objective function: "singular" for NCT
targeted to one single effect; "composite" for NCT targeted to multiple
effects; "penalized" for a OF computed while considering a single effect only
and including a penalization term related to the variance.}

\item{output}{Desired output of the analysis. if output = "detection" only
point estimates are computed, if output = "estimation" both estimated effects
and variances are computed}
}
\value{
A Network Causal Tree - NCT - object. An NCT object is a data frame reporting the
results of the Network Causal Trees, where each tree is characterized by a set of partitions.
Specifically the GOF variable includes the value of the GOF in the corresponding partition, the NOBS_TR column
includes the number of obs belonging to the training set and to the given partition,
the column FILTER includes the values of the Xs that identify the given partition, the
IDTREE variable identifies the id of the tree ,  the
NUMTREE variable represents the number of trees that identify the given partition,
the NOBS_EST column includes the number of obs belonging
to the estimation set and to the given partition, the columns EFF1000_EST, EFF1101_EST,  EFF1110_EST,
EFF0100_EST reporting the estimated effects in the partitions. If output = "Estimation" the data frame
also include the columns SETAU1000_EST, SETAU1101_EST,  SETAU1110_EST, SETAU0100_EST reporting the estimated standard errors of the effects in the partitions
}
\description{
Returns a Network Causal Tree, with the corresponding estimates
}
