% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_NCT}
\alias{plot_NCT}
\title{Plot of a Network Causal Tree}
\usage{
plot_NCT(
  NCT,
  varnames,
  title,
  output = "estimation",
  coloreff = "1000",
  vcolor = c("seagreen4", "seagreen1", "lightblue1", "dodgerblue2"),
  vsize = 32,
  vsize2 = 32,
  vshape = "circle",
  vframecolor = "black",
  vlabelcolor = "black",
  vlabelcex = 0.8,
  vlabelfont = 1,
  elabelcex = 0.7,
  elabelfamily = "sans",
  ecolor = "black",
  ewidth = 0.3,
  edge.arrow.size = 0.5,
  elabelcolor = "black",
  ot = FALSE,
  colleg = "black",
  coltitleleg = "black",
  font.main = 1,
  cex.main = 1,
  adj = 1,
  col.main = "black"
)
}
\arguments{
\item{NCT}{NetworkCausalTrees results.}

\item{varnames}{(Ordered) covariates names vector}

\item{title}{Title}

\item{output}{Output of the NCT. if output = "detection" only point estimates
are reported, if output = "estimation" both estimated effects and variances are reported}

\item{coloreff}{Estimated Effect according to which nodes are colored
It can be "1000", "0100","1110","1101"}

\item{vcolor}{String Vector of four colors: the first two identify the
extremes of the color palette employed to depict negative values, while the
second two identify the extremes of the color palette employed to depict
positive values. Note that the intensity of the color is related to the Statistic of the
estimated effect in each leaf (Default values represent negative estimates
with greens and positive estimates with blues).}

\item{vsize}{Vertex size}

\item{vsize2}{Vertex size - to be set if you want rectangular vertices -}

\item{vshape}{Vertex shape}

\item{vframecolor}{Vertex's frame color}

\item{vlabelcolor}{Vertex label color
Title}

\item{vlabelcex}{Vertex Label cex}

\item{vlabelfont}{Vertex label font}

\item{elabelcex}{Edge's label cex}

\item{elabelfamily}{Edge label family}

\item{ecolor}{Edges color}

\item{ewidth}{Edge width}

\item{edge.arrow.size}{Edge's arrow size}

\item{elabelcolor}{Edges' label color
Vertices}

\item{ot}{TRUE if the NCT object includes an optimal treatment evaluation,
FALSE otherwise (default: FALSE)
Edges}

\item{colleg}{Legend color}

\item{coltitleleg}{Legend's title color}

\item{cex.main}{Title cex}

\item{adj}{adj}

\item{col.main}{Title color
Legend}
}
\description{
Plots the Network Causal Tree results
General
}
