% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_NCT}
\alias{plot_NCT}
\title{Plot of a Network Causal Tree}
\usage{
plot_NCT(
  NCT,
  cov_names,
  title,
  output = "estimation",
  effect_color_nodes = "1000",
  vertex_color = c("lightskyblue", "dodgerblue2"),
  vertex_size = 44,
  vertex_size2 = 40,
  vertex_shape = "rectangle",
  vertex_frame_color = "black",
  vertex_label_color = "black",
  vertex_label_cex = 2,
  vertex_label_font = 1,
  edge_label_cex = 1.5,
  edge_label_family = "sans",
  edge_color = "black",
  edge_width = 0.3,
  edge_arrow_size = 0.5,
  edge_label_color = "black",
  legend_color = "black",
  legend_title_color = "black",
  main_font = 1,
  main_cex = 1,
  adj = 1,
  main_color = "black"
)
}
\arguments{
\item{NCT}{A NCT object (output of the NetworkCausalTree function).}

\item{cov_names}{(Ordered) covariates names vector.}

\item{title}{Title.}

\item{output}{If "detection" only point estimates are reported, if
"estimation" both estimated effects and variances are reported.
Edge - specific parameters}

\item{effect_color_nodes}{The estimated effect according to which nodes are colored
It can be "1000", "0100","1110","1101".}

\item{vertex_color}{String Vector of two colors identifying the
extremes of the color palette employed to depict positive
estimates of the effect specified in "effect_color_nodes"
(Negative values are represented in white )}

\item{vertex_size}{Vertex size.}

\item{vertex_size2}{Vertex size (for rectangular vertices).}

\item{vertex_shape}{Vertex shape.}

\item{vertex_frame_color}{Vertex's frame color.}

\item{vertex_label_color}{Vertex label color.
Title - specific parameters}

\item{vertex_label_cex}{Vertex Label cex.}

\item{vertex_label_font}{Vertex label font.}

\item{edge_label_cex}{Edge's label cex.}

\item{edge_label_family}{Edge label family.}

\item{edge_color}{Edges color.}

\item{edge_arrow_size}{Edge's arrow size.}

\item{edge_label_color}{Edges' label color.
Vertices - specific parameters}

\item{legend_color}{Legend color.}

\item{legend_title_color}{Legend's title color.}

\item{main_font}{Title font.

Legend - specific parameters}

\item{main_cex}{Title cex.}

\item{adj}{adj.}

\item{main_color}{Title color.}

\item{ewidth}{Edge width.}
}
\value{
A plot representing the estimated Network causal tree.
Each node reports - from the top to the bottom
\itemize{
\item i) the estimated effect 10000
ii) the estimated effect 0100
iii) the size of the sub popuedgeson
Colors provide an intuition on the strength of the effect specified in "effect_color_nodes".
}
}
\description{
Plots the Network Causal Tree results
General parameters
}
