% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vartau1000.R
\name{Vartau1000}
\alias{Vartau1000}
\title{Estimated Variance for the effect 1000}
\usage{
Vartau1000(N, Y, W, G, p, Ne, Ne_list)
}
\arguments{
\item{N}{Sample size}

\item{Y}{N x 1 vector, Observed Outcome}

\item{W}{N x 1 vector, Individual Treatment}

\item{G}{N x 1 vector, Neighborhood Treatment}

\item{p}{N x 1 vector, Probability to be assigned to the active individual intervention}

\item{Ne}{N x 1 vector, Degree}

\item{Ne_list}{List of N elements - where N is the sample size -
where each element i contains the IDs of the direct neighbors of unit i}
}
\value{
A numeric value corresponding to the estimated variance of the effect
of interest.
}
\description{
Computes the estimated variance of the effect 1000. This effect
compares individuals with the joint treatment exposure (w=1,g=0)
with those with (w=0,g=0)
}
